1. update dependency; 2. dao -> repo
