+++ app/build.gradle
@@ -31,7 +31,7 @@ apply plugin: 'com.ofg.uptodate'
 apply plugin: 'com.getkeepsafe.dexcount'
 apply from: '../AndroidCodeQualityConfig/quality.gradle'
 apply plugin: 'com.bugtags.library.plugin'
-apply plugin: 'android-git-sha'
+//apply plugin: 'android-git-sha'
 apply plugin: 'net.ltgt.errorprone'
 
 /*def keyConfigPath
@@ -87,7 +87,6 @@ android {
     dexOptions {
         // Skip pre-dexing when running on Travis CI or when disabled via -Dpre-dex=false.
         preDexLibraries = preDexEnabled && !isTravis
-        incremental true
         maxProcessCount 8
         javaMaxHeapSize = isTravis ? "2g" : "6g"
     }
@@ -160,6 +159,7 @@ android {
 
 unMock {
     downloadFrom 'https://oss.sonatype.org/content/groups/public/org/robolectric/android-all/6.0.0_r1-robolectric-0/android-all-6.0.0_r1-robolectric-0.jar'
+    downloadTo "${System.getenv('HOME')}/.gradle/caches/"
 
     keep "android.text.TextUtils"
 }
@@ -171,11 +171,11 @@ if (!isTravis) {
     }
 }
 
-gitSha {
-    buildTypeMatcher = 'release'
-    flavorMatcher = '.*'
-    abortIfGitDirty = false
-}
+//gitSha {
+//    buildTypeMatcher = 'release'
+//    flavorMatcher = '.*'
+//    abortIfGitDirty = false
+//}
 
 dependencies {
     def appDependencies = rootProject.ext.appDependencies
