+++ build.gradle
@@ -5,6 +5,7 @@ buildscript {
     repositories {
         jcenter()
         mavenCentral()
+        mavenLocal()
         maven {
             // For fabric
             url 'https://maven.fabric.io/public'
@@ -21,13 +22,13 @@ buildscript {
         classpath 'com.getkeepsafe.dexcount:dexcount-gradle-plugin:0.5.4'
         classpath 'de.mobilej.unmock:UnMockPlugin:0.5.0'
         classpath 'com.ofg:uptodate-gradle-plugin:1.6.2'
-        classpath 'com.github.okbuilds:okbuild-gradle-plugin:0.1.7'
+        classpath 'com.github.okbuilds:okbuild-gradle-plugin:0.2.1'
         classpath "net.rdrei.android.buildtimetracker:gradle-plugin:0.8.0"
         classpath 'com.frogermcs.androiddevmetrics:androiddevmetrics-plugin:0.4'
         classpath 'com.bugtags.library:bugtags-gradle:1.2.6'
         classpath 'com.github.promeg:android-git-sha-plugin:1.0.0'
         classpath "net.ltgt.gradle:gradle-errorprone-plugin:0.0.8"
-        classpath 'com.squareup.sqldelight:gradle-plugin:0.4.2'
+        classpath 'com.squareup.sqldelight:gradle-plugin:0.4.3'
     }
     configurations.classpath.exclude group: 'com.android.tools.external.lombok'
 }
@@ -47,6 +48,9 @@ allprojects {
         maven { url "https://jitpack.io" }
         maven {
             url  "http://dl.bintray.com/piasy/maven"
+        }
+        maven {
+            url  "http://dl.bintray.com/okbuild/maven"
         }
     }
 }
