+++ buildsystem/dependencies.gradle
@@ -1,10 +1,10 @@
 ext {
     // Android
-    androidBuildToolsVersion = '23.0.1'
+    androidBuildToolsVersion = '24.0.1'
     androidMinSdkVersion = 15
-    androidTargetSdkVersion = 23
-    androidCompileSdkVersion = 23
-    androidSupportSdkVersion = '23.3.0'
+    androidTargetSdkVersion = 24
+    androidCompileSdkVersion = 24
+    androidSupportSdkVersion = '24.0.0'
     versionCode = 1
     versionName = '1.0.0'
 
@@ -14,11 +14,12 @@ ext {
     daggerVersion = '2.5'
     butterKnifeVersion = '8.1.0'
 
-    okHttpVersion = '3.3.1'
+    okHttpVersion = '3.4.1'
     retrofitVersion = '2.1.0'
     rxBindingVersion = '0.4.0'
     rxLifeCycleVersion = '0.6.1'
 
+    sqlDelightVersion = '0.4.3'
     autoBundleVersion = '1.0.2'
     onActivityResultVersion = '0.6.0'
 
@@ -35,12 +36,12 @@ ext {
 
     baseDependencies =
             [androidSupportAppCompatV7: "com.android.support:appcompat-v7:${androidSupportSdkVersion}",
-             safelyAndroid            : "com.github.piasy:safelyandroid:1.1.1",
+             safelyAndroid            : "com.github.piasy:safelyandroid:1.2.3",
 
              javaxAnnotation          : "org.glassfish:javax.annotation:10.0-b28",
 
              mvp                      : "com.hannesdorfmann.mosby:mvp:2.0.1",
-             rxJava                   : "io.reactivex:rxjava:1.1.6",
+             rxJava                   : "io.reactivex:rxjava:1.1.7",
              rxAndroid                : "io.reactivex:rxandroid:1.2.1",
              rxBinding                : "com.jakewharton.rxbinding:rxbinding:${rxBindingVersion}",
              rxBindingAppCompatV7     : "com.jakewharton.rxbinding:rxbinding-appcompat-v7:${rxBindingVersion}",
@@ -55,7 +56,8 @@ ext {
              retrofit                 : "com.squareup.retrofit2:retrofit:${retrofitVersion}",
              retrofitGsonConverter    : "com.squareup.retrofit2:converter-gson:${retrofitVersion}",
              retrofitRxJavaAdapter    : "com.squareup.retrofit2:adapter-rxjava:${retrofitVersion}",
-             sqlBrite                 : 'com.squareup.sqlbrite:sqlbrite:0.6.3',
+             sqlBrite                 : 'com.squareup.sqlbrite:sqlbrite:0.7.0',
+             sqlDelightRuntime        : "com.squareup.sqldelight:runtime:${sqlDelightVersion}",
 
              threetenabp              : "com.jakewharton.threetenabp:threetenabp:1.0.3",
              autoBundle               : "com.github.yatatsu:autobundle:${autoBundleVersion}",
@@ -73,7 +75,7 @@ ext {
             [recyclerViewV7     : "com.android.support:recyclerview-v7:${androidSupportSdkVersion}",
              cardViewV7         : "com.android.support:cardview-v7:${androidSupportSdkVersion}",
 
-             rxJavaProGuardRules: "com.artemzin.rxjava:proguard-rules:1.1.6.0",
+             rxJavaProGuardRules: "com.artemzin.rxjava:proguard-rules:1.1.7.0",
 
              once               : "com.jonathanfinerty.once:once:1.1.0",
 
@@ -85,7 +87,7 @@ ext {
              SSIV               : "com.davemorrissey.labs:subsampling-scale-image-view:3.5.0",
              slidr              : "com.r0adkll:slidableactivity:2.0.5",
              recyclerViewPager  : "com.github.lsjwzh.RecyclerViewPager:lib:v1.1.0-beta5",
-             statusBarUtil      : "com.jaeger.statusbaruitl:library:1.1.0",
+             statusBarUtil      : "com.jaeger.statusbaruitl:library:1.2.0",
 
              bugTags            : "com.bugtags.library:bugtags-lib:1.2.7",
 
@@ -93,7 +95,7 @@ ext {
              stetho             : "com.facebook.stetho:stetho:${stethoVersion}",
              strictmodeNotifier : "com.nshmura:strictmode-notifier:0.9.1",
              blockCanary        : "com.github.moduth:blockcanary-android:1.2.1",
-             anrWatchDog        : "com.github.anrwatchdog:anrwatchdog:1.2.0",
+             anrWatchDog        : "com.github.anrwatchdog:anrwatchdog:1.3.0",
              leakCanary         : "com.squareup.leakcanary:leakcanary-android:1.4-beta2",]
 
     aptDependencies = [daggerApt          : "com.google.dagger:dagger-compiler:${daggerVersion}",
@@ -105,8 +107,9 @@ ext {
                        autoBundleApt      : "com.github.yatatsu:autobundle-processor:${autoBundleVersion}",
                        onActivityResultApt: "com.vanniktech:onactivityresult-compiler:${onActivityResultVersion}",
 
+                       sqlDelightBin      : "com.github.okbuilds:SqlDelightBin:${sqlDelightVersion}",
                        retrolambda        : "net.orfjackal.retrolambda:retrolambda:2.3.0",
-                       errorprone         : "com.google.errorprone:error_prone_core:2.0.9",]
+                       errorprone         : "com.google.errorprone:error_prone_core:2.0.11",]
 
     testDependencies = [junit                  : "junit:junit:4.12",
                         mockito                : "org.mockito:mockito-core:2.0.71-beta",
